requirements.txt
asgiref==3.5.0
Django==4.0.3
django-taggit==2.1.0
djangorestframework==3.13.1
gunicorn==20.1.0
Markdown==3.3.6
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1

python-dotenv~=0.20.0
pytest~=7.1.2